Handle links to external domains
There are two cases when dealing with links pointing to an external
domain. Either already the link itself contains a foreign domain or
the link leads to a forward because of a HTTP 3xx status code.

Due to the fact, that also a seemingly foreign domain link might
redirect to the target domain a HTTP HEAD request is performed in
any case. This way the final location can be asserted.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>