Add graph data structure to organize data
The natural data structure for this domain is a graph. Each page and
each static asset can be represented as a node. If there is an edge
between two nodes they are interlinked.

In order to make the lookup as easy as possible a hash is used where the
URI of a page or static asset is used to hash the node itself.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>