Model sitemap and use SLIM for template rendering
The resulting sitemap will be represented by its own object. This class
is also used to perform the template rendering. I have added SLIM to be
used as the template engine.

A vendor directory will contain third-party JavaScript libraries and
stylesheets. For now it contains Bootstrap, jQuery and d3.js.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>